+++ src/lib.rs
@@ -7,4 +7,3 @@ pub mod unicodesegmenters;
 pub mod lexer;
 pub mod parser;
 pub mod symbols;
-pub mod environment;